Add support for hb-debug-console PLDM BIOS attribute
The primary object of this commit is to provide an interface that
will read the hb-debug-console attribute from the BMC and translate
it to a boolean value. While adding this support I realized a lot
of the code could be shared between this and the code read reading
the hb-hyp-switch attribute Hostboot already supports reading so
some refactoring has been done.

Change-Id: I69f2260406631c3647d712db1dd121a6ad9379a0
RTC: 208803
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108586
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>