Support presence detection of targets w/ remotely sourced VPD
For VPD such as the Platform VPD, the host does not always have
physical access to the device and must get the VPD from another
entity such as the BMC. In this case the host will not be able to
do a simple eeprom read as we typically do for a presence detection
test. In order to check the presence status we will set the
VPD_SWITCH bit vpdCollectedRemotely when we have collected the
VPD for a remote device on each boot and that is the bit we will check
when users call pvpdPresent to check if pvpd is present or not.

Change-Id: I9d7cb7c78bf081bf60b2bd1c11e02c10308f24e6
CQ: SW512826
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108653
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>