Add error handling and callouts for unsupported SBE PSU operations
This commit adds extra error handling for unsupported SBE PSU
operations.

If we encounter an unsupported SBE PSU operation on the primary/master
processor and the error will halt the IPL, we can fix the problem by
performing an immediate SBE update.

If we encounter such an error on a subsidiary processor, then the best
we can do is deconfigure it, because we can't update the SBE before
the SMP fabric is up (since FSI access to the SBE SEEPROMs are blocked
for security reasons), and after the SMP fabric is up we we
immediately update the subsidiary SBEs anyway.

In any case we add a callout for the SBE code to the error returned
from performPsuChipOp.

Change-Id: Ie159ee0d6759860fbfe4149a2533923cb70cc1ea
RTC: 173219
CMVC-Prereq: 1120964
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108706
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>