Added API getVtocRecordMetaData to retrieve the VTOC record meta data
There was duplicate code that returned the VTOC record data and I was
about to add another.  So I created a single method to do this but
added an optional argument that can expedite the extraction of the
VTOC meta data.

This extra argument will be used in a future commit.  For now this is
just a consolidation of code and a removal of redundant code.

Also added some more constants to get away from "magic numbers" and
to consolidate constant literal strings.

Change-Id: Ie0dc1d1e7ee898d420974a46ef1a3b5e1a563e58
RTC: 250100
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108812
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>