Remove presence check in SMP Deconfig rules
There is a check for presence on SMPA type IOHS busses that is not
necessary in _invokeDeconfigAssocProc. It was supposed to mirror a P9
like check for P10 type targets but it does so erroneously. In P9, ABUS
connections needed to be present for SMP deconfig logic to be applied to
them. With the transition to IOHS busses, SMPA type busses were being
assumed to be the "out-group" type bus that ABUSses are in P9 but that
cannot be assumed and must be verified via GROUP ID.

Change-Id: I6efd20653cc9479888fe691f05fc09ff15d9f152
RTC:246169
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108911
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>