Print link register in printk backtraces
This commit adds support for printing the link register (LR) out in
printk backtraces. This is useful because the current backtrace code
only prints return addresses that have been saved to the stack; but
small leaf functions (like memcpy) often never save their return
address to the stack, and instead just keep it in the LR for
efficiency. If a fault happens within such a function, we can't tell
where it was called from without printing the LR.

Change-Id: I7dd8ca75f52df8c415b28e7c0f239668e540a1d2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109105
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>