Rebuild EECACHE when version mismatch is detected
At the start of the EECACHE section of pnor there is a version field
that tells the user what layout version this EECACHE is. If we encounter
a version that does not match the LATEST version the code supports, we
should clear the EECACHE section of pnor and trigger a reconfig loop
in order to build it again with the latest version of the EECACHE layout.

Change-Id: Iee8f1a6c4eef573caaa68cc6067d80055ff6faa3
RTC: 247228
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109173
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>