4-byte align FRU table response buffer
In the response to our PLDM::getFruTableMetaData request, the table
size reported by the BMC does not include the padding bytes in order
to force 4-byte alignment. As a result, Hostboot was not allocating
enough bytes in the response buffer. The openBMC APIs we are using
are copying however many bytes are available in the response message.
This was causing an array overrun when we had fru_table_len %4 != 0.

Change-Id: I2df14bbd761aad36878fe057ca7c0256759c3473
CQ: SW513927
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109180
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Andrew J Geissler <andrewg@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>