Enable Error Logging for EECACHE
Modify asserts to log errors to make
visible the problems encountered.

EECACHE is now masking out the upper byte
of the HUID (which indicates the NODE ID).

By masking out the NODE ID on the EECACHE entry
enables the EECACHE to be node-agnostic.

The NODE ID was discovered to cause the maximum
number of EECACHE records to be exceeded prematurely
when NODE's may have been physically moved from
one drawer slot position to another drawer slot
position (which dicates the NODE ID value).

A new attribute is also introduced which allows
an auto-recovery when reaching the maximum EECACHE
entry limit.

By default when the maximum EECACHE limit is reached,
we will auto-recover by clearing the EECACHE, enabling
a dynamic reset (which will be a visible notification
in the error logs).

If desired in lab or debug scenarios this new attribute
can be set to disable the auto-recovery, making testing
or debug more accommodating.

Change-Id: Ic0d6a98802af9ab79bba3363c15be9774f7f1703
CQ:SW512018
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109271
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>