Send tracelite output to VUART2 when hb-debug-console is "Enabled"
There is a PLDM BIOS attribute called hb-debug-console which can
be set on the BMC and it tells us if we should write debug output
to VUART2, which is Hostboot's debug console during IPL time. If
this attribute is set to "Enabled" then we will write tracelite
content to VUART2 which can be parsed by the weave tool.

Change-Id: Ib1a60e97ea0e090734132482ca58dc78a83ae68d
RTC: 208803
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109348
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>