Code to update the ECC data when a record has been modified
This is the code that will update the ECC data of a record
when the record has been modified.

This commit also has the supporting API's to retrieve the meta data
of the record and the API to check the results from the update call.

Dropped the ECC update TODO within the IpVpdFacade::write method because
it is being doing done in the mvpdWrite.  The mvpdWrite is an API that
will call the IpVpdFacade::write.  Putting the ECC update within the
IpVpdFacade::write will require to check every record to be certain that
it is an MVPD record, this could be time consuming.

Change-Id: I82cba9b04853521f6987da3a2c1c8839ec806c5e
RTC: 250100
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109303
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>