Create fake gard records for deconfigured parts
In order for Hostboot to communicate semi-persistent
deconfigurations down to the BMC, Hostboot needs to
create pseudo gard records in the PNOR.  This is the
same thing we do on FSP systems, except the producer
is different.  On FSP, HWSV creates deconfiguration
records based on the deconfiguration requests HB sends
down in the error logs and Hostboot then consumes them
when it starts up fresh on the reconfig reboot.  For BMC
systems, the data will be consumed by Hostboot in the same
way, but also by the BMC for any of the units on the master processor.

Change-Id: Ie12f59d46e710ad9a8e3283e3df8aebae54b0ff1
RTC:255318
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106873
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>