Remove Finding HW Keys' Hash and Secure Version from end of HBBL Section
This commit removes finding the HW Keys' Hash and the Secure Version
from the end of the HBBL section in the Hostboot bootloader.  As a
temporary measure we are passing in the imprint/developer HW Keys' Hash
and a Secure Version of Zero.  There will be a follow-on commit to
remove this temporary measure once the SBE code passes these values
up to the Hostboot bootloader.

Change-Id: I2272f2eb2f32f145c80dddd1b032096babd9a434
RTC:208821
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109459
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>