Increase the error log printk buffer capture size
We've recently added some information to the printk backtrace which
can take up a significant amount of space. Error logs only collect the
most recent text in the printk buffer, so it's possible that when the
backtrace info is large, important information is too far back in the
buffer to be captured in an error log.

This commit increases the printk buffer capture size for error logs by
about 800 bytes to 2k, so that it's more likely that important
debugging information is captured in the printk output.

Change-Id: Ib7a7af91b1b25be86396150b1b1b6e238cfbbdb0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109587
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>