Force early devtree sync if hardware has changed
Normally we will only sync our attributes down to the devtree
at the end of step16.  This leaves us with a mismatched set of
data on the BMC for most of the IPL on a genesis boot or when
parts are added and removed.  To improve that, we will do a
sync right after we do presence detection so that we have an
accurate representation of the present parts as soon as possible.

Change-Id: Ib1e0949c60151426743225763daaa0ec9e6a6bd2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109801
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>