Refclock handling: Add mode for optional redundancy
Add new REFCLOCK_SELECT modes for degraded operation, where we
want to boot in redundant mode but don't trust one of the
oscillators any more, so we try to set up redundant mode but don't
fail if the secondary oscillator fails a test.

While we're in there, factor the clock test logic out into a
shared function, and implement code to transfer the attribute
value into the SBE via the scratch registers.

Change-Id: Ia0c990e3a2d985084e1c0c5078f22bea579f8ce7
Original-Change-Id: I2babd8b02311c902f839b60ee311391e7eb1cc90
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109824
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Manish Kumar Chowdhary <manichow@in.ibm.com>
Reviewed-by: ARAVIND T. NAIR <aravindnair@in.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122534
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>