Fix pmic_enable 4U bug due to stale GPIO cfg register
Since we now clear individual active-low bits in the GPIO
config register, we assumed and asserted the other bits
would remain unchanged, and this resulted in a logic bug.
On FW, bits seem to mangle between power cycles, so this
workaround ensures that the bits are reset properly, in
order to avoid incorrectly declaring N-Mode.

Change-Id: I9fcb18144cbcde9c32dc5f39ab42d464f3f87165
CQ: SW515055
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/109889
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>