Increase the max number of in/out group connections for SMP logic
In SMP wrap testing it was discovered that there was an array overrun of
the max number of connections for in/out group IOHS links. This commit
increases the max to account for SMP Wrap mode and adds asserts to
ensure that overruns are detected easier/faster in the future.

Change-Id: I944f8a9f0887ee5dc7e2966222cec1b1e68db564
CQ:FW708977
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110061
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>