HB CI hb-trace wrapped logs
Name the hb-trace log file to reflect that the
content is derived using the hb-trace tools to aide the log
consumer to help differentiate the various log(s) content.

For example the hbTracMERG file under the Build Artifacts output
during HB CI contains the complete trace output (typical
TRACFCOMP) and the TS_INFO/TS_FAIL output which is collected
when the post-IPL CXXTESTs are executed.

The hb-trace wrapped output will contain up-to-the-latest
output (possibly wrapped), whereas the hbTracMERG output may
not have flushed to the file at the point of possible
failures, etc.

Change-Id: I67707ae093020cdc72b8a73e7cb3bb212d798ab8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110087
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>