FCO: Changes to processing field core override code and multicast handling code
Relocated the method that processes the field core override,
DeconfigGard::processFieldCoreOverride, from the HWAS common code to
the HWAS PLAT code and renamed to DeconfigGard::platProcessFieldCoreOverride.

This has been moved to the PLAT area because HB will be running this code
and will be a no-op for HWSV.

Updated the getting parent code to handle getting the parent of a multicast
CORE target.

Change-Id: I7f09e530edea3f0779537fca287fd03b289ee524
RTC:260384
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110100
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>