Use POSITION instead of ORDINAL_ID in FRU RSID encoding (commit 1/n)
This commit changes the PLDM FRU RSID encoder to use the POSITION
attribute, rather than the ORDINAL_ID attribute, to encode which
instance of a target an RSID refers to.

This commit is needed because ORDINAL_ID can be larger than the number
of bits we have allocated for its encoding. It does not always start
at zero and is not always contiguous, so its numeric value is inflated
(for DIMM targets in particular). POSITION does start at zero and is
densely packed, so that we don't overflow our encoding schema.

Change-Id: I93974f2fd7f1ffce9954e640bb51c5c1d81e7e47
RTC: 254920
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110198
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>