VPD EECACHE Synchronization
Add refreshVPD to the flow in cacheEECACHEPartition
(typically this is performed in the MPIPL path).

During MPIPL we need to query ALL targets which hold the
EECACHE_VPD_STATE attribute.  If a target has the
EECACHE_VPD_STATE set to VPD_NEEDS_REFRESH, then we will
force a synchronization between the PNOR EECACHE partition
and the eecache in memory/seeproms.

This situation occurs when HBRT does a VPD write which is
-NOT- persisted to the PNOR EECACHE partition.

Change-Id: Ie02e02076a447f24f7bedd88a7f356f55d61313c
RTC:251157
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110305
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>