Everest support: Process split SMP link targets (commit 1/n)
This commit adds support for processing split SMP link targets in the
Everest MRW using the "new" SMP layout (where SMPGROUP targets are
located under XBUS or ABUS targets, which are under IOHS
targets). Eventually all the MRWs will move to this style, but this
commit retains backward-compatibility with the old style for Rainier
and Denali.

Output for the Rainier and Denali MRWs is not affected, save for
setting the CHIPLET_ID of the SMPGROUP targets to match their parent
IOHSes.

Change-Id: I5e5269d8e7a2ffa3cbd541ce544b741d7de9fa75
RTC: 264654
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110413
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>