Add rt_sbe interface to override the in-memory copy of EECACHE at runtime
This commit adds a new passThruCmd interface in rt_sbe.C which will allow
overwrites to the in-memory copy of EECACHE at runtime. Updates through
this new interface will only update the in-memory copy of EECACHE used at
runtime, and will not update the EECACHE copy in PNOR nor the original
copy of individual EECACHE entries on the respective HW devices.

Change-Id: Ie8c53f0053661a7ed46092968cc1ad82a4a905b1
RTC: 260834
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110582
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>