Attribute Sync Presence Detection
Need to initially -NOT- allow attribute synchronization until
after host_discover_targets has completed.

During EEPROM cache building we may encounter the need to CLEAR
the EECACHE and then perform a RECONFIG loop.  If the attributes
are synchronized down to the Service Processor this leads to
the primary processor being -NOT- identified as functional.  We
need to set the capability to perform attribute synchronizations
based on being POST host_discover_targets.

Change-Id: Ibb1ee8d18d9b1eac39cb348bddf6bb3dd4ffc9de
CQ:SW516525
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110642
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>