Set iv_gard and iv_deconfig flags on errlEntry during errl commit
Previously we would set these flags on the creation of a hw, clock
or part callout if the callout had a gard or deconfig associated
with it. We desire more flexibility in changing the gard/deconfig
setting for a callout after it is created and reversing these flags
can be tricky so we will defer setting these flags until we have
committed the error log because at that point the gard/deconfig
setting for the callouts will not change.

Change-Id: I60153238dc52e2a0a24fb11a5ccce032fadd23ac
RTC: 244857
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110676
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>