Add new APIs to errlEntry for manipulating existing callouts
There are some instances where we want to alter the gard/deconfig
settings associated with a hw, clock, or part callout. This commit
adds 3 APIs. One which allows the user to query the callouts on an
error log looking for a callout for a given target. The other two
APIs allow the user to set the deconfig or gard elements in all
callouts of an error log for a given target.

Change-Id: I4c8a09072cf9a486ee3e734d4e0d475ed2a6af0d
RTC: 244857
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110685
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>