Expanded the error log within method getVtocRecordMetaData to be more explicit
The error log that is produced in the method getVtocRecordMetaData when
there is an issue when we retrieve the VTOC record covers 2 error issues and
is vague about which error the log is for.

This became an issue when I had a defect that produced this error and I had
to spend time digging into the code to determine exactly what the issue was.

This code change will break out these two error conditions and be more explicit
about what the error is. This will expedite and facilitate a HB investigation
into such error without having to spend time digging into the code.

Change-Id: If4645025622fcd7e7aff0111e823fbfe576fb5ef
CQ: SW516717
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110719
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>