Cache devtree nodes for target lookup
This commit adds functionality to cache the physical path -> node
mapping for device tree synchronization.

The libfdt function fdt_node_offset_by_prop_value, which is used to
look up the device tree node associated with a target from its
physical path attribute, iterates (potentially) the entire device tree
(every attribute on every node) to locate the node corresponding to
that physical path's target. This function is called once for every
target in the system blueprint, and so it consumes a considerable
amount of execution time.

Rather than iterate the entire device tree for every target, this
commit creates an in-memory mapping from target handle to device tree
node, using a single iteration of the device tree. This reduces the
complexity of lookup by nearly an order of magnitude, and improves IPL
speeds.

Change-Id: Ie1f5a3e0ba9b31dfa129d69aba66ec0cbe5e64f4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111023
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>