Fix terrible istep performance in secure mode
Due to Hostboot failing to associate the OVERRIDES_ATTEMPTED_FLAG
to the system target, the thread that handles istep attribute
synching crashes in a secure boot only path when it attempts to get
that attribute.  This commit reassociates the attribute to the system
target in the HB only target XML file.

Change-Id: I20fcc2ceddea274458411260ca070f4043075054
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111147
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>