Handle deconfigured proc0 on FSPB boot
The basic issue is that we always boot from the memory range owned
by topology id 0.  Typically this is owned by proc0.  In this
case, proc0 is deconfigured though so we can't use it.  Hostboot
has logic to swap topoology ids if we have no functional memory
behind proc0, but the code wasn't correctly handling the case
where the processor itself is deconfigured.

Change-Id: I5304ddef32e52221d93533b439cec74ed04dabec
CQ: SW516583
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111179
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>