Enable backup TPM intialization
Enable the call to initBackupTpm in host_secureboot_lockdown, which
is istep 10.3, now that this is working. In addition to enabling this
ensure that we are not attempting to do the intialization if the TPM
is not functional. Also skip the log replay from primary->backup that
was done in previous generation as the TPMs are now independent. In
fact we will go a step further and always poison the backup TPM after
initialization to ensure the TPMs are disjoint.

Change-Id: I4565594bf416ef7fd6376baf555c5c45d1d4c425
RTC: 264774
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111180
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>