Force termination when attribute override task crashes
If the attribute override task (entry point @ platTaskEntry)
crashes for any reason, Hostboot will not terminate since the task
is not detached and therefore must be waited on by the primary task
which isn't going to do that.  This commit detaches the attribute
override task so that when it crashes, Hostboot kernel will detect
the problem and force a termination.

Change-Id: I598847468755ea0a0776e8616016c1c18c1315ee
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111198
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>