Add HBD_RW PNOR Partitions
Enable the HBD PNOR partition to support HPT
to allow the capability to unpin the designated
readonly HBD payload part to free up memory
constraints in early IPL.

The determination of where to map the RW parts
of the HBD data is logically determined by the
presence of the HPT on the HBD PNOR partition.

During the building of the PNOR the HBD will either
be marked with HPT or without HPT.  The logic will then
switch based on this flag to optionally bring in the
HBD_RW PNOR partition and map this RW section for the
Attribute Resource Provider.  The XML to build the PNOR
is the basis for enabling the HBD_RW section to be built,
which later will get exploited dynamically during IPL.

op-build will exploit this feature in future work.

Change-Id: Ie9dd29da44b02a278d3eef55422be0c703273392
RTC:266901
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111297
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>